+++ src/main/java/lotto/view/ResultView.java
@@ -1,4 +1,8 @@
-package lotto;
+package lotto.view;
+
+import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
+import lotto.domain.Rank;
 
 import java.util.HashMap;
 import java.util.List;
